Extracting common precompiled contract behavior from BridgePerformanceTest and BridgePerformanceTestCase into common framework
- New PrecompiledContractPerfomanceTest class
- New PrecompiledContractPerformanceTestCase class
- BridgePerformanceTest now extends PrecompiledContractPerformanceTest
- BridgePerformanceTestCase now extends PrecompiledContractPerformanceTestCase
