+++ rskj-core/src/test/java/co/rsk/peg/performance/BridgePerformanceTestCase.java
@@ -21,7 +21,6 @@ package co.rsk.peg.performance;
 import co.rsk.bitcoinj.core.*;
 import co.rsk.config.BridgeConstants;
 import co.rsk.config.RskSystemProperties;
-import co.rsk.config.TestSystemProperties;
 import co.rsk.db.RepositoryImpl;
 import co.rsk.db.RepositoryTrackWithBenchmarking;
 import co.rsk.db.TrieStorePoolOnMemory;
@@ -29,138 +28,31 @@ import co.rsk.peg.Bridge;
 import co.rsk.peg.BridgeStorageConfiguration;
 import co.rsk.peg.BridgeStorageProvider;
 import co.rsk.test.builders.BlockChainBuilder;
-import co.rsk.trie.TrieImpl;
-import co.rsk.vm.VMPerformanceTest;
-import org.ethereum.config.blockchain.regtest.RegTestGenesisConfig;
 import org.ethereum.core.Blockchain;
 import org.ethereum.core.Repository;
 import org.ethereum.core.Transaction;
 import org.ethereum.crypto.ECKey;
 import org.ethereum.crypto.HashUtil;
-import org.ethereum.datasource.HashMapDB;
 import org.ethereum.vm.LogInfo;
 import org.ethereum.vm.PrecompiledContracts;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.BeforeClass;
-import org.bouncycastle.util.encoders.Hex;
 
-import java.lang.management.ManagementFactory;
-import java.lang.management.ThreadMXBean;
 import java.math.BigInteger;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.stream.Collectors;
 
-public abstract class BridgePerformanceTestCase {
+public abstract class BridgePerformanceTestCase extends PrecompiledContractPerformanceTestCase {
     protected static NetworkParameters networkParameters;
     protected static BridgeConstants bridgeConstants;
-    private static TestSystemProperties config;
-
-    private boolean oldCpuTimeEnabled;
-    private ThreadMXBean thread;
-
-    protected class ExecutionTracker {
-        private static final long MILLION = 1_000_000;
-
-        private final ThreadMXBean thread;
-        private long startTime, endTime;
-        private long startRealTime, endRealTime;
-        private RepositoryTrackWithBenchmarking.Statistics repositoryStatistics;
-
-        public ExecutionTracker(ThreadMXBean thread) {
-            this.thread = thread;
-        }
-
-        public void startTimer() {
-            // Everything is expressed in nanoseconds
-            startTime = thread.getCurrentThreadCpuTime();
-            startRealTime = System.currentTimeMillis() * MILLION;
-        }
-
-        public void endTimer() {
-            endTime = thread.getCurrentThreadCpuTime();
-            endRealTime = System.currentTimeMillis() * MILLION;
-        }
-
-        public void setRepositoryStatistics(RepositoryTrackWithBenchmarking.Statistics statistics) {
-            this.repositoryStatistics = statistics;
-        }
-
-        public RepositoryTrackWithBenchmarking.Statistics getRepositoryStatistics() {
-            return this.repositoryStatistics;
-        }
-
-        public long getExecutionTime() {
-            return endTime - startTime;
-        }
-
-        public long getRealExecutionTime() {
-            return endRealTime - startRealTime;
-        }
-    }
 
     @BeforeClass
-    public static void setUpBeforeClass() throws Exception {
-        config = new TestSystemProperties();
-        config.setBlockchainConfig(new RegTestGenesisConfig());
+    public static void setup() throws Exception {
         bridgeConstants = config.getBlockchainConfig().getCommonConstants().getBridgeConstants();
         networkParameters = bridgeConstants.getBtcParams();
     }
 
-    @AfterClass
-    public static void printStatsIfNotInSuite() throws Exception {
-        if (!BridgePerformanceTest.isRunning()) {
-            BridgePerformanceTest.printStats();
-        }
-    }
-
-    @Before
-    public void setupCpuTime() {
-        thread = ManagementFactory.getThreadMXBean();
-        if (!thread.isThreadCpuTimeSupported()) {
-            throw new RuntimeException("Thread CPU time not supported");
-        };
-
-        oldCpuTimeEnabled = thread.isThreadCpuTimeEnabled();
-        thread.setThreadCpuTimeEnabled(true);
-    }
-
-    @After
-    public void teardownCpuTime() {
-        thread.setThreadCpuTimeEnabled(oldCpuTimeEnabled);
-    }
-
-    @After
-    public void forceGC() {
-        long sm = Runtime.getRuntime().freeMemory();
-        VMPerformanceTest.forceGc();
-        long em = Runtime.getRuntime().freeMemory();
-        System.out.println(String.format("GC - free mem before: %d, after: %d", sm, em));
-    }
-
-    protected static class Helper {
-        public static Transaction buildTx(ECKey sender) {
-            return buildSendValueTx(sender, BigInteger.ZERO);
-        }
-
-        public static Transaction buildSendValueTx(ECKey sender, BigInteger value) {
-            byte[] gasPrice = Hex.decode("00");
-            byte[] gasLimit = Hex.decode("00");
-
-            Transaction tx = new Transaction(
-                    null,
-                    gasPrice,
-                    gasLimit,
-                    sender.getAddress(),
-                    value.toByteArray(),
-                    null);
-            tx.sign(sender.getPrivKeyBytes());
-
-            return tx;
-        }
-
+    protected static class Helper extends PrecompiledContractPerformanceTestCase.Helper {
         // Federates for bridge regtest constants
         public static final List<ECKey> FEDERATOR_ECKEYS = Arrays.stream(new String[]{
                 "federator1",
@@ -172,20 +64,8 @@ public abstract class BridgePerformanceTestCase {
             return Helper.FEDERATOR_ECKEYS.get(Helper.randomInRange(0, Helper.FEDERATOR_ECKEYS.size()-1));
         }
 
-        public static int randomInRange(int min, int max) {
-            return new Random().nextInt(max - min + 1) + min;
-        }
-
         public static Coin randomCoin(Coin base, int min, int max) {
             return base.multiply(randomInRange(min, max));
-        }
-
-        public static HeightProvider getRandomHeightProvider(int max) {
-            return (int executionIndex) -> new Random().nextInt(max);
-        }
-
-        public static HeightProvider getRandomHeightProvider(int min, int max) {
-            return (int executionIndex) -> randomInRange(min, max);
         }
 
         public static BtcBlock generateAndAddBlocks(BtcBlockChain btcBlockChain, int blocksToGenerate) {
@@ -235,10 +115,6 @@ public abstract class BridgePerformanceTestCase {
             return (int executionIndex) -> Helper.buildTx(new ECKey());
         }
 
-        public static TxBuilder getZeroValueTxBuilder(ECKey sender) {
-            return (int executionIndex) -> Helper.buildTx(sender);
-        }
-
         public static BridgeStorageProviderInitializer buildNoopInitializer() {
             return (BridgeStorageProvider provider, Repository repository, int executionIndex) -> {};
         }
@@ -246,74 +122,6 @@ public abstract class BridgePerformanceTestCase {
 
     protected interface BridgeStorageProviderInitializer {
         void initialize(BridgeStorageProvider provider, Repository repository, int executionIndex);
-    }
-
-    protected interface TxBuilder {
-        Transaction build(int executionIndex);
-    }
-
-    protected interface ABIEncoder {
-        byte[] encode(int executionIndex);
-    }
-
-    protected interface HeightProvider {
-        int getHeight(int executionIndex);
-    }
-
-    private ExecutionTracker execute(
-            ABIEncoder abiEncoder,
-            BridgeStorageProviderInitializer storageInitializer,
-            TxBuilder txBuilder, HeightProvider heightProvider, int executionIndex) {
-
-        ExecutionTracker executionInfo = new ExecutionTracker(thread);
-
-        RepositoryImpl repository = createRepositoryImpl(config);
-        Repository track = repository.startTracking();
-        BridgeStorageConfiguration bridgeStorageConfigurationAtThisHeight = BridgeStorageConfiguration.fromBlockchainConfig(config.getBlockchainConfig().getConfigForBlock(executionIndex));
-        BridgeStorageProvider storageProvider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstants,bridgeStorageConfigurationAtThisHeight);
-
-        storageInitializer.initialize(storageProvider, track, executionIndex);
-
-        try {
-            storageProvider.save();
-        } catch (Exception e) {
-            throw new RuntimeException("Error trying to save the storage after initialization", e);
-        }
-        track.commit();
-
-        Transaction tx = txBuilder.build(executionIndex);
-
-        List<LogInfo> logs = new ArrayList<>();
-
-        RepositoryTrackWithBenchmarking benchmarkerTrack = new RepositoryTrackWithBenchmarking(repository);
-
-        Bridge bridge = new Bridge(config, PrecompiledContracts.BRIDGE_ADDR);
-        Blockchain blockchain = BlockChainBuilder.ofSize(heightProvider.getHeight(executionIndex));
-        bridge.init(
-                tx,
-                blockchain.getBestBlock(),
-                benchmarkerTrack,
-                blockchain.getBlockStore(),
-                null,
-                logs
-        );
-
-        // Execute a random method so that bridge support initialization
-        // does its initial writes to the repo for e.g. genesis block,
-        // federation, etc, etc. and we don't get
-        // those recorded in the actual execution.
-        bridge.execute(Bridge.GET_FEDERATION_SIZE.encode());
-        benchmarkerTrack.getStatistics().clear();
-
-        executionInfo.startTimer();
-        bridge.execute(abiEncoder.encode(executionIndex));
-        executionInfo.endTimer();
-
-        benchmarkerTrack.commit();
-
-        executionInfo.setRepositoryStatistics(benchmarkerTrack.getStatistics());
-
-        return executionInfo;
     }
 
     protected ExecutionStats executeAndAverage(String name,
@@ -324,23 +132,66 @@ public abstract class BridgePerformanceTestCase {
              HeightProvider heightProvider,
              ExecutionStats stats) {
 
-        long sm = Runtime.getRuntime().freeMemory();
+        EnvironmentBuilder environmentBuilder = new EnvironmentBuilder() {
+            private Bridge bridge;
+            private RepositoryTrackWithBenchmarking benchmarkerTrack;
 
-        for (int i = 0; i < times; i++) {
-            System.out.println(String.format("%s %d/%d", name, i+1, times));
+            private RepositoryImpl createRepositoryImpl(RskSystemProperties config) {
+                return new RepositoryImpl(null, new TrieStorePoolOnMemory(), config.detailsInMemoryStorageLimit());
+            }
 
-            ExecutionTracker tracker = execute(abiEncoder, storageInitializer, txBuilder, heightProvider, i);
+            @Override
+            public Environment initialize(int executionIndex) {
+                RepositoryImpl repository = createRepositoryImpl(config);
+                Repository track = repository.startTracking();
+                BridgeStorageConfiguration bridgeStorageConfigurationAtThisHeight = BridgeStorageConfiguration.fromBlockchainConfig(config.getBlockchainConfig().getConfigForBlock(executionIndex));
+                BridgeStorageProvider storageProvider = new BridgeStorageProvider(track, PrecompiledContracts.BRIDGE_ADDR, bridgeConstants,bridgeStorageConfigurationAtThisHeight);
 
-            stats.executionTimes.add(tracker.getExecutionTime());
-            stats.realExecutionTimes.add(tracker.getRealExecutionTime());
-            stats.slotsWritten.add(tracker.getRepositoryStatistics().getSlotsWritten());
-            stats.slotsCleared.add(tracker.getRepositoryStatistics().getSlotsCleared());
-        }
+                storageInitializer.initialize(storageProvider, track, executionIndex);
 
-        return stats;
-    }
+                try {
+                    storageProvider.save();
+                } catch (Exception e) {
+                    throw new RuntimeException("Error trying to save the storage after initialization", e);
+                }
+                track.commit();
 
-    public static RepositoryImpl createRepositoryImpl(RskSystemProperties config) {
-        return new RepositoryImpl(new TrieImpl(null, true), new HashMapDB(), new TrieStorePoolOnMemory(), config.detailsInMemoryStorageLimit());
+                Transaction tx = txBuilder.build(executionIndex);
+
+                List<LogInfo> logs = new ArrayList<>();
+
+                benchmarkerTrack = new RepositoryTrackWithBenchmarking(repository, new TrieStorePoolOnMemory(), config.detailsInMemoryStorageLimit());
+
+                bridge = new Bridge(config, PrecompiledContracts.BRIDGE_ADDR);
+                Blockchain blockchain = BlockChainBuilder.ofSizeWithNoTransactionPoolCleaner(heightProvider.getHeight(executionIndex));
+                bridge.init(
+                        tx,
+                        blockchain.getBestBlock(),
+                        benchmarkerTrack,
+                        blockchain.getBlockStore(),
+                        null,
+                        logs
+                );
+
+                // Execute a random method so that bridge support initialization
+                // does its initial writes to the repo for e.g. genesis block,
+                // federation, etc, etc. and we don't get
+                // those recorded in the actual execution.
+                bridge.execute(Bridge.GET_FEDERATION_SIZE.encode());
+                benchmarkerTrack.getStatistics().clear();
+
+                return new Environment(
+                        bridge,
+                        benchmarkerTrack
+                );
+            }
+
+            @Override
+            public void teardown() {
+                benchmarkerTrack.commit();
+            }
+        };
+
+        return super.executeAndAverage(name, times, environmentBuilder, abiEncoder, txBuilder, heightProvider, stats);
     }
 }
